fix(augroups): Remove events enum (#85)
* CI: update inputs via PR

* fix(augroups): Remove events enum

Apparently not all events are documented in the same place (I thought
they where all
[here](https://neovim.io/doc/user/autocmd.html#autocmd-events). And you
can even add new events? See lazy.nvim's
[User Events](https://github.com/folke/lazy.nvim#-user-events).

This makes the current events enum overly restrictive. I came across
this when trying to use the `LspAttach`, and although I could add it, it
doesn't scale if plugins can add their own events. At most we could fire
a warning if it is not in the enum, but that would get annoying pretty
quick.

---------

Co-authored-by: Joscha Loos <j.loos@posteo.net>